requirements.txt
# 教学文档自动生成系统 - 依赖清单

# --- 核心依赖 ---
pyyaml>=6.0
openai>=1.30.0

# --- LLM SDK（按需安装，非必须全装）---
anthropic>=0.30.0       # Anthropic Claude API
google-genai>=1.0.0     # Google Gemini API

# --- 信息采集 ---
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
trafilatura>=1.8.0      # 网页正文提取
arxiv>=2.1.0            # arXiv 论文检索
feedparser>=6.0.0       # RSS 订阅
duckduckgo-search>=6.0.0  # DuckDuckGo 搜索

# --- 向量数据库 ---
chromadb>=0.5.0
sentence-transformers>=3.0.0

# --- 文档导出 ---
python-docx>=1.1.0      # Word 导出
markdown>=3.6           # Markdown 转 HTML
reportlab>=4.2.0        # 出版级 PDF 渲染

# --- 可选：PDF 导出（需要系统级依赖）---
# weasyprint>=62.0       # 取消注释以启用 PDF 导出

# --- 去重 ---
rapidfuzz>=3.0.0
